+++ packages/slate-react/src/components/editor.js
@@ -3,7 +3,7 @@ import Portal from 'react-portal'
 import React from 'react'
 import SlateTypes from 'slate-prop-types'
 import Types from 'prop-types'
-import logger from 'slate-dev-logger'
+import warning from 'slate-dev-warning'
 import { Schema, Stack } from 'slate'
 import memoizeOne from 'memoize-one'
 
@@ -136,11 +136,10 @@ class Editor extends React.Component {
 
     // If we've resolved a few times already, and it's exactly in line with
     // the updates, then warn the user that they may be doing something wrong.
-    if (resolves > 5 && resolves === updates) {
-      logger.warn(
-        'A Slate <Editor> component is re-resolving `props.plugins` or `props.schema` on each update, which leads to poor performance. This is often due to passing in a new `schema` or `plugins` prop with each render by declaring them inline in your render function. Do not do this!'
-      )
-    }
+    warning(
+      resolves > 5 && resolves === updates,
+      'A Slate <Editor> component is re-resolving `props.plugins` or `props.schema` on each update, which leads to poor performance. This is often due to passing in a new `schema` or `plugins` prop with each render by declaring them inline in your render function. Do not do this!'
+    )
 
     if (change) {
       this.onChange(change)
@@ -231,7 +230,8 @@ class Editor extends React.Component {
 
   change = (...args) => {
     if (this.tmp.isChanging) {
-      logger.warn(
+      warning(
+        true,
         "The `editor.change` method was called from within an existing `editor.change` callback. This is not allowed, and often due to calling `editor.change` directly from a plugin's event handler which is unnecessary."
       )