+++ packages/slate/src/index.js
@@ -24,7 +24,6 @@ import Stack from './models/stack'
 import Text from './models/text'
 import TextUtils from './utils/text-utils'
 import Value from './models/value'
-import { resetKeyGenerator, setKeyGenerator } from './utils/generate-key'
 import { resetMemoization, useMemoization } from './utils/memoize'
 
 /**
@@ -51,11 +50,9 @@ export {
   PathUtils,
   Point,
   Range,
-  resetKeyGenerator,
   resetMemoization,
   Schema,
   Selection,
-  setKeyGenerator,
   Stack,
   Text,
   TextUtils,
@@ -80,11 +77,9 @@ export default {
   PathUtils,
   Point,
   Range,
-  resetKeyGenerator,
   resetMemoization,
   Schema,
   Selection,
-  setKeyGenerator,
   Stack,
   Text,
   TextUtils,