+++ packages/slate/src/models/node.js
@@ -1,5 +1,5 @@
 import isPlainObject from 'is-plain-object'
-import logger from 'slate-dev-logger'
+import warning from 'slate-dev-warning'
 import { List } from 'immutable'
 
 import Block from './block'
@@ -32,9 +32,9 @@ class Node {
       let { object } = attrs
 
       if (!object && attrs.kind) {
-        logger.deprecate(
-          'slate@0.32.0',
-          'The `kind` property of Slate objects has been renamed to `object`.'
+        warning(
+          true,
+          'As of slate@0.32.0, the `kind` property of Slate objects has been renamed to `object`.'
         )
 
         object = attrs.kind
@@ -121,9 +121,9 @@ class Node {
     let { object } = value
 
     if (!object && value.kind) {
-      logger.deprecate(
-        'slate@0.32.0',
-        'The `kind` property of Slate objects has been renamed to `object`.'
+      warning(
+        true,
+        'As of slate@0.32.0, the `kind` property of Slate objects has been renamed to `object`.'
       )
 
       object = value.kind