+++ Changelog.md
@@ -4,7 +4,7 @@ Since Slate is a monorepo with many packages that are versioned separately, we m
 
 * [`slate`](./packages/slate/Changelog.md)
 * [`slate-base64-serializer`](./packages/slate-base64-serializer/Changelog.md)
-* [`slate-dev-logger`](./packages/slate-dev-logger/Changelog.md)
+* [`slate-dev-warning`](./packages/slate-dev-warning/Changelog.md)
 * [`slate-html-serializer`](./packages/slate-html-serializer/Changelog.md)
 * [`slate-hyperscript`](./packages/slate-hyperscript/Changelog.md)
 * [`slate-plain-serializer`](./packages/slate-plain-serializer/Changelog.md)